REF: Extract dispatch logic
