+++ src/components/ImageGridCard.tsx
@@ -1,4 +1,4 @@
-import { dispatchRequestArgs } from "@/state/ZuState";
+import { dispatchRequestArgs } from "@/state/dispatchHelpers";
 import { truncateText } from "@/utils/dataUtils";
 import { TypeItem } from "@/utils/types";
 import { memo, Ref } from "react";
