+++ src/components/SearchToolBarDropdown.tsx
@@ -1,4 +1,5 @@
-import { dispatchRequestArgs, useStore } from "@/state/ZuState";
+import { useStore } from "@/state/ZuState";
+import { dispatchRequestArgs } from "@/state/dispatchHelpers";
 import { capitalize } from "@/utils/dataUtils";
 import { TypeState } from "@/utils/types";
 import { FormEvent } from "react";
